% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{add_map_labels}
\alias{add_map_labels}
\title{Add labels to points on map}
\usage{
add_map_labels(map, data, labels, legendGroup = NULL, ...)
}
\arguments{
\item{map}{leaflet map}

\item{data}{track data}

\item{...}{}
}
\description{
Add labels to points on map
}
