% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anemometer1}
\alias{anemometer1}
\title{System matrices for 1-parameter anemometer.}
\format{
A list:
\describe{
\item{E}{a diagonal mass matrix}
\item{A1, A2}{sparse coefficient matrices}
\item{B}{input vector}
\item{C}{output (sparse) vector}
\item{P}{a permutation matrix, maps nodes to DOFs, exported from ANSYS}
}
}
\source{
\url{http://modelreduction.org/index.php/Anemometer}
}
\usage{
anemometer1
}
\description{
System dimension 29,008, with one input and one output.
}
\keyword{datasets}
