% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anemometer3}
\alias{anemometer3}
\title{System matrices for 3-parameter anemometer.}
\format{
A list:
\describe{
\item{Es, Ef}{diagonal mass matrices}
\item{Ads, Ac}{symmetric sparse coefficient matrices}
\item{Adf}{asymmetric sparse coefficient matrix}
\item{B}{input vector, different from 1-parameter anemometer}
\item{C}{output (sparse) vector, same with 1-parameter anemometer}
}
}
\source{
\url{http://modelreduction.org/index.php/Anemometer}
}
\usage{
anemometer3
}
\description{
System dimension 29,008, with one input and one output.
}
\keyword{datasets}
