% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{AngleVS}
\alias{AngleVS}
\title{Angle between a vector and a subspace}
\usage{
AngleVS(v, X)
}
\arguments{
\item{X}{an orthonormal frame of a subspace, an n-by-k matrix}
}
\value{
a scalar, \link{0, pi/2}
}
\description{
Angle between a vector and a subspace
}
