% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{backwardEuler}
\alias{backwardEuler}
\title{backward Euler (aka implicit Euler) scheme
Solves ODE: \eqn{E \dot{x} = A x + b}}
\usage{
backwardEuler(E, A, b, dt = timestep, nt = steps)
}
\arguments{
\item{b}{stationary input vector}
}
\value{
a matrix of state vector snapshots
}
\description{
backward Euler (aka implicit Euler) scheme
Solves ODE: \eqn{E \dot{x} = A x + b}
}
