% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{base_map}
\alias{base_map}
\title{Create a blank map}
\usage{
base_map(providers = NULL, labels = NULL)
}
\arguments{
\item{providers}{character vector of tile providers}

\item{labels}{character vector of provider labels}
}
\value{
map
}
\description{
Create a blank map
}
