% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{blockHadamard}
\alias{blockHadamard}
\title{Compute the Hadamard product of XtX and kronecker(S, 1_k 1_k^T),
without computing the Kronecker product.}
\usage{
blockHadamard(XtX, S)
}
\description{
Compute the Hadamard product of XtX and kronecker(S, 1_k 1_k^T),
without computing the Kronecker product.
}
\note{
Much slower than XtX * kronecker(DKDinv, Matrix(1, k, k)).
}
