% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_API.R
\name{coord2tz}
\alias{coord2tz}
\title{Get timezone from lat/lon coordinates via Google Timezone API}
\usage{
coord2tz(lat, lon, api_key)
}
\arguments{
\item{lat}{latitude (double)}

\item{lon}{longitude (double)}

\item{api_key}{Google Timezone API key}
}
\value{
timezone ID (character vector)
}
\description{
Get timezone from lat/lon coordinates via Google Timezone API
}
