% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation_gps.R
\name{creation_gps}
\alias{creation_gps}
\title{Permet de creer les positions et le temps de quatre satellites}
\usage{
creation_gps(data, num, num_satellites = 4, arrondi = FALSE)
}
\arguments{
\item{data}{Une tibble contenant les positions des endroits sur terre. Le tibble contient le \code{nom},
la \code{latitude} en degres, la \code{longitude} en degres et l'\code{altitude} en metres.}

\item{num_satellites}{Le nombre de satellites a creer. La valeur par defaut est 4.}

\item{arrondi}{Si \code{TRUE}, nous trouvons des positions de satellites entieres. Si \code{FALSE}, nous trouvons des positions
decimales plus proches de la realite.}

\item{k}{Le numero de l'endroit dont on veut trouver la position GPS.}
}
\value{
sat_final Une tibble dont les lignes representent les satellites et les colonnes les coordonnees spatiales et temporelles.
  En coordonnees cartesiennes x, y et z, ainsi qu'en temps t.
}
\description{
Permet de creer les positions et le temps de quatre satellites
}
