% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation_problemes.R
\name{creation_problemes}
\alias{creation_problemes}
\title{Permet de creer deux documents Rmd pour les etudiants et les enseignants}
\usage{
creation_problemes(N, database, n_sat = 4, arrondi = FALSE)
}
\arguments{
\item{N}{Un nombre entier indiquant le nombre total de problemes a creer.}

\item{database}{Une tibble contenant les endroits voulus.}

\item{n_sat}{Le nombre de satellites a creer. La valeur par defaut est 4.}

\item{arrondi}{Si \code{TRUE}, nous trouvons des positions de satellites entieres. Si \code{FALSE}, nous trouvons des positions
decimales plus proches de la realite.}
}
\value{
Deux documents Rmd
}
\description{
Permet de creer deux documents Rmd pour les etudiants et les enseignants
}
