% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel-SE.R
\name{defaultLength}
\alias{defaultLength}
\title{Rule-of-thumb lengthscale of the SE kernel for GP-subspace}
\usage{
defaultLength(d, l)
}
\arguments{
\item{d}{Parameter dimension}

\item{l}{Sample size}
}
\description{
Rule-of-thumb lengthscale of the SE kernel for GP-subspace
}
