% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROM.R
\name{directROM}
\alias{directROM}
\title{Compute reduced system matrices given local reduced bases}
\usage{
directROM(p, listA, listV, E, B, C)
}
\arguments{
\item{p}{a vector of parameters used for computing the reduced bases}

\item{listA}{a list of coefficient matrices, from \code{getA()}.}

\item{listV}{a list of local reduced bases}
}
\value{
a list: an array of reduced standardized coefficient matrices,
matrices of reduce standardized input-to-state vectors and state-to-output vectors.
}
\description{
Compute reduced system matrices given local reduced bases
}
\note{
Require global objects (for 1-parameter anemometer): E, B, c.
}
