% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_speed.R
\name{dist_from_start}
\alias{dist_from_start}
\title{Distance from first to last trackpoint (straight line)}
\usage{
dist_from_start(track)
}
\arguments{
\item{track}{}
}
\value{
dbl meters
}
\description{
Distance from first to last trackpoint (straight line)
}
