% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_speed.R
\name{dist_to_finish}
\alias{dist_to_finish}
\title{Distance between current position and finish}
\usage{
dist_to_finish(track, route)
}
\arguments{
\item{route}{}
}
\value{
dbl meters
}
\description{
Distance between current position and finish
}
