% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distvec}
\alias{distvec}
\title{Distance vector from a point to a set of points}
\usage{
distvec(x, x0)
}
\arguments{
\item{x}{a set of points, either a length-n vector or an n-by-p matrix}

\item{x0}{a point, either a scalar or a length-p vector}
}
\description{
Distance vector from a point to a set of points
}
