% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_speed.R
\name{extract_coord}
\alias{extract_coord}
\title{Extract coordinates from route or track}
\usage{
extract_coord(track, func)
}
\arguments{
\item{track}{data.frame}

\item{func}{function, e.g., min, max, mean}
}
\value{
c(lat, lon)
}
\description{
Extract coordinates from route or track
}
