% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{GapMetric}
\alias{GapMetric}
\title{The gap metric between two subspaces: sine of the largest principal angle}
\usage{
GapMetric(X, Y)
}
\arguments{
\item{X, }{Y Steifel representations.}
}
\description{
The gap metric between two subspaces: sine of the largest principal angle
}
