% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolate-Subspace.R
\name{GL2i}
\alias{GL2i}
\title{Compute Grassmann logarithm at p_i for all points p_j, j != il.}
\usage{
GL2i(i)
}
\description{
Compute Grassmann logarithm at p_i for all points p_j, j != il.
}
\note{
Require (listXtrain)
}
