% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSR-predict.R
\name{GPSubspacePredEVD}
\alias{GPSubspacePredEVD}
\title{GP subspace regression prediction, eigen-decomposition version
36ms, w/o messaging (k=20,l=12); faster than SVD version!}
\usage{
GPSubspacePredEVD(thetaNew, thetaTrain, len, t = k)
}
\description{
GP subspace regression prediction, eigen-decomposition version
36ms, w/o messaging (k=20,l=12); faster than SVD version!
}
