% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSR-predict-svd.R
\name{GPSubspacePredSVD}
\alias{GPSubspacePredSVD}
\title{44ms, w/o messaging (k=20,l=12).
GP subspace regression prediction, posterior from equal likelihood}
\usage{
GPSubspacePredSVD(thetaNew, thetaTrain, len, t = k, method = "rsvd")
}
\arguments{
\item{thetaNew}{a target parameter point}

\item{t}{truncation level, defaults to k; if NULL, use thin SVD.}
}
\value{
a list: Vcirc, principal directions as coordinates in the global basis,
a c-by-t matrix (more than the mean prediction);
sigma2, principal variances, a vector of length t;
eps2, noise variance, a scalar.
}
\description{
44ms, w/o messaging (k=20,l=12).
GP subspace regression prediction, posterior from equal likelihood
}
\details{
To get the explicit principal directions, do matrix multiplication V = stdX$u \%*\% Vcirc.
}
\note{
Require (K, XtX, VbtX)
}
