% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSR-preprocess.R
\name{GPSubspacePrepQR}
\alias{GPSubspacePrepQR}
\title{Preprocessing via Householder QR with column pivoting}
\usage{
GPSubspacePrepQR(X, transposeR = TRUE)
}
\arguments{
\item{X}{Concatenated orthonormal bases.}

\item{transposeR}{Whether to transpose the R matrix,
which saves the transpose in solving linear equations in prediction.}
}
\value{
A list:
Pt, transpose of the column pivot matrix.
}
\description{
Preprocessing via Householder QR with column pivoting
}
\note{
R's \code{base::qr()} uses LINPACK's \code{DQRDC} for Householder QR with column pivoting.
The C++ library Eigen has \code{ColPivHouseholderQR}, which may be faster
and can be accessed via \code{RcppEigen}.
}
