% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSR-preprocess.R
\name{GPSubspacePrepSVD}
\alias{GPSubspacePrepSVD}
\title{Preprocessing via thin SVD}
\usage{
GPSubspacePrepSVD(X)
}
\arguments{
\item{X}{Concatenated orthonormal bases.}
}
\description{
Preprocessing via thin SVD
}
