% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_parse.R
\name{gpx_creator}
\alias{gpx_creator}
\title{Parse creator of a GPX file}
\usage{
gpx_creator(gpxFile)
}
\arguments{
\item{gpxFile}{File path of the file to be parsed}
}
\value{

}
\description{
Parse creator of a GPX file
}
