% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSR-LOOCV-error.R
\name{gradLOODist2Evd}
\alias{gradLOODist2Evd}
\title{Gradient of squared LOOCV Riemannian distance}
\usage{
gradLOODist2Evd(i, Kinv, rGKinv, VbtX, trunk = 2 * k)
}
\arguments{
\item{rGKinv}{an order-l matrix that consists of \eqn{\partial k_{p,q} / k_{p,q}}.}

\item{trunk}{Extended truncated EVD for approximate pseudo-inverse, defaults to 2*k.}
}
\description{
Gradient of squared LOOCV Riemannian distance
}
\note{
Require: (XtX; k, l, Jk)
}
