% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel-SE.R
\name{gradSEmat}
\alias{gradSEmat}
\title{Gradient of the squared exponential kernel, matrix version}
\usage{
gradSEmat(x, len = 1, K = NULL)
}
\arguments{
\item{K}{Kernel matrix, if already computed}
}
\description{
Gradient of the squared exponential kernel, matrix version
}
