% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{GrassmannExp}
\alias{GrassmannExp}
\title{Grassmann exponential}
\usage{
GrassmannExp(X0, T)
}
\arguments{
\item{T}{horizontal lift of a tangent vector at X0.}
}
\value{
a Stiefel representation of exp_\href{T}{X0}
}
\description{
Grassmann exponential
}
