% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{GrassmannLog}
\alias{GrassmannLog}
\title{Grassmann logarithm}
\usage{
GrassmannLog(X0, Y)
}
\arguments{
\item{X0}{Stiefel representation of base point on the Grassmann manifold}

\item{Y}{Stiefel representation of target point}
}
\value{
the horizontal lift of log_\link{X0} \link{Y}
}
\description{
Grassmann logarithm
}
\references{
@Bendokat2020
}
