% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSR-LOOCV-error.R
\name{gSSDist}
\alias{gSSDist}
\title{Gradient of sum of squared LOOCV Riemannian distances}
\usage{
gSSDist(len)
}
\description{
Gradient of sum of squared LOOCV Riemannian distances
}
\note{
require: (XtX; k, Jk, l)
}
