% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSR-LOOCV-pdf.R
\name{h}
\alias{h}
\title{Criterion to minimize, the negative log LOO predictive probability.}
\usage{
h(len)
}
\description{
Criterion to minimize, the negative log LOO predictive probability.
}
\note{
require global (thetaTrain, XtX)
}
