% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_speed.R
\name{haversine}
\alias{haversine}
\title{Distance between two points (haversine formula)}
\usage{
haversine(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lon2}{}
}
\value{
meters
}
\description{
Distance between two points (haversine formula)
}
