% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSR-LOOCV-error.R
\name{hSSDist}
\alias{hSSDist}
\title{Criterion to minimize, the LOOCV prediction error: sum of squared Riemannian distances.}
\usage{
hSSDist(len)
}
\description{
Criterion to minimize, the LOOCV prediction error: sum of squared Riemannian distances.
}
\note{
require (thetaTrain, XtX, VbtX; getKinv)
}
