% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel-SE.R
\name{kerSEmat}
\alias{kerSEmat}
\title{Squared exponential kernel}
\usage{
kerSEmat(x, x0 = NULL, len = 1)
}
\arguments{
\item{x}{Vector of scalar parameters, or matrix of vector parameters.}

\item{x0}{New parameter. If provided, computes a covariance vector.}

\item{len}{Length-scale of correlation, defaults to 1.}

\item{x1, x2}{Input parameters, real numbers}
}
\value{
If only x is provided, a \code{dpoMatrix}
(positive-semidefinite matrix from \code{Matrix}, with Cholesky factor computed).
If both are provided, a vector of length nrow(x), if matrix; or length(x), if vector.
}
\description{
Squared exponential kernel
}
