% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{LagrangeInterp}
\alias{LagrangeInterp}
\title{A function factory for Lagrange interpolation.}
\usage{
LagrangeInterp(p)
}
\arguments{
\item{p}{Interpolation points}
}
\value{
a function that takes in a new point and
returns the weight vector for Lagrange interpolation.
}
\description{
A function factory for Lagrange interpolation.
}
