% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSR-LOOCV-pdf.R
\name{LogDetCov}
\alias{LogDetCov}
\title{Log determinant of predictive "covariance matrix"
\eqn{log(det(\tilde{\Sigma}))}}
\usage{
LogDetCov(n, sigma2, eps2)
}
\description{
Log determinant of predictive "covariance matrix"
\eqn{log(det(\tilde{\Sigma}))}
}
