% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSR-LOOCV-pdf.R
\name{LogDetCrossBox}
\alias{LogDetCrossBox}
\title{Log determinant of the principal submatrices of breveBoxinv}
\usage{
LogDetCrossBox(breveBoxinv)
}
\arguments{
\item{breveBoxinv}{inverse of breveBox}
}
\value{
a vector of length l, one for each \eqn{log(det(crossBox_i))}.
}
\description{
\eqn{log(det(crossBox_i)) = -log(det(X_i^T \tilde{\Sigma}_{-i}^{-1} X_i))}
}
\note{
Require (k)
}
