% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSR-LOOCV-pdf.R
\name{LogDetLooCov}
\alias{LogDetLooCov}
\title{Log determinant of leave-one-out predictive "covariance matrix"
\eqn{log(det(\tilde{\Sigma}_{-i}))}
Result does not match raw computation!}
\usage{
LogDetLooCov(i)
}
\description{
Log determinant of leave-one-out predictive "covariance matrix"
\eqn{log(det(\tilde{\Sigma}_{-i}))}
Result does not match raw computation!
}
\note{
require (X, XtX; Kinv)
}
