% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSR-LOOCV-pdf.R
\name{LogDetXSinvX}
\alias{LogDetXSinvX}
\title{Log determinant of "inverse match matrix"
\eqn{log(det(X^T \tilde{\Sigma}^{-1} X))}}
\usage{
LogDetXSinvX(X, V, sigma2, eps2)
}
\description{
Log determinant of "inverse match matrix"
\eqn{log(det(X^T \tilde{\Sigma}^{-1} X))}
}
