% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSR-LOOCV-pdf.R
\name{LogLooPredProb}
\alias{LogLooPredProb}
\title{Leave-one-out predictive probability density
Faster than using \code{LogDetLooCov()} and \code{LogDetCrossBox()}.}
\usage{
LogLooPredProb(i, Kinv, logdetbreveBox, logdetcrossBoxi)
}
\arguments{
\item{i}{index of left out parameter point}

\item{Kinv}{inverse of correlation matrix}

\item{logdetbreveBox}{log det of breveBox}

\item{logdetcrossBoxi}{log det of crossBox_i, computed from \code{LogDetCrossBox()}.}
}
\description{
Leave-one-out predictive probability density
Faster than using \code{LogDetLooCov()} and \code{LogDetCrossBox()}.
}
\note{
Require global (XtX)
}
