% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_gps}
\alias{map_gps}
\title{Display a GPS track and/or route on a map}
\usage{
map_gps(track = NULL, route = NULL, track_labels = NULL, save = FALSE)
}
\arguments{
\item{save}{}
}
\value{
leaflet map
}
\description{
Display a GPS track and/or route on a map
}
