% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_path}
\alias{map_path}
\title{Display series of coordinates on a map}
\usage{
map_path(map, data, labels = NULL, circleColor = "#FF4900",
  circleRadius = 50, connectPoints = TRUE, lineColor = "#FF4900",
  lineType = NULL, noHide = F, legendGroup = NULL)
}
\arguments{
\item{noHide}{}
}
\value{
a leaflet map
}
\description{
Display series of coordinates on a map
}

