% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_path_lines}
\alias{map_path_lines}
\title{Add path lines to map}
\usage{
map_path_lines(map, data, legendGroup = NULL, lineColor = "#FF4900", ...)
}
\arguments{
\item{map}{leaflet map}

\item{data}{track data}

\item{...}{}
}
\value{
map
}
\description{
Add path lines to map
}
