% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_path_points}
\alias{map_path_points}
\title{Add path points to map}
\usage{
map_path_points(map, data, legendGroup = NULL, circleColor = "#FF4900",
  circleRadius = 5, ...)
}
\arguments{
\item{map}{leaflet map}

\item{data}{track data}

\item{...}{}
}
\value{
map
}
\description{
Add path points to map
}
