% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_track}
\alias{map_track}
\title{Create a map of a track}
\usage{
map_track(track, freq = "60 min", tz_offset = 0, providers = NULL,
  labels = NULL, circleColor = NULL, lineColor = NULL)
}
\arguments{
\item{track}{track data}

\item{freq}{frequency for trk_reduce}

\item{tz_offset}{hours offset from GMT}

\item{lineColor}{}
}
\description{
Create a map of a track
}
