% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merveilles-data.R
\docType{data}
\name{merveilles}
\alias{merveilles}
\title{Une tibble contenant des donnees geodetiques pour GPS}
\format{Un objet de classe \code{"tibble"}}
\usage{
data(merveilles)
}
\description{
Les localisations geodetiques des merveilles utilisees
pour construire les problemes GPS. Les informations sont
le \code{nom} de la merveille, la \code{latitude} de la
merveille, la \code{longitude} de la merveille et
son \code{altitude}.
}
\examples{
data(merveilles)
noms <- merveilles$nom
longitudes <- merveilles$longitude
}
\keyword{datasets}
