% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{MultiquadricRBF}
\alias{MultiquadricRBF}
\title{Hardy's multiquadrics, a radial basis function}
\usage{
MultiquadricRBF(p, f)
}
\arguments{
\item{p}{Interpolation points, a n-by-d matrix}

\item{f}{Function values, a length-n vector or an n-by-q matrix (vector-valued function)}
}
\value{
A function that takes in a new point
and returns the value of the interpolation function.
}
\description{
Hardy's multiquadrics, a radial basis function
}
\note{
Parameter computed per \link{@Amsallem2010, Appendix H}
}
