% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_speed_time.R
\name{p_track_elap_HM}
\alias{p_track_elap_HM}
\title{Print elapsed time of track}
\usage{
p_track_elap_HM(track)
}
\arguments{
\item{track}{}
}
\value{
String - e.g., 12:20
}
\description{
Displays in Hours:Minutes
}

