% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{PiecewiseLinearInterp}
\alias{PiecewiseLinearInterp}
\title{A function factory for piecewise linear interpolation.}
\usage{
PiecewiseLinearInterp(p)
}
\arguments{
\item{p}{Interpolation points, can be unordered.}
}
\value{
a function that takes in a new point and
returns the weight vector for Lagrange interpolation.
}
\description{
A function factory for piecewise linear interpolation.
}
