% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{PrAngles}
\alias{PrAngles}
\title{Principal angles between two subspaces}
\usage{
PrAngles(X, Y)
}
\value{
a vector of principal angles in increasing order, \link{0, pi/2}
}
\description{
Principal angles between two subspaces
}
