% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_API.R
\name{process_inreach_xml}
\alias{process_inreach_xml}
\title{Process inReach XML API response}
\usage{
process_inreach_xml(xml, attrib)
}
\arguments{
\item{attrib}{}
}
\description{
Process inReach XML API response
}
