% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_progress.R
\name{progress_per_time}
\alias{progress_per_time}
\title{Get progress per time interval}
\usage{
progress_per_time(track, finish_lat, finish_lon, freq)
}
\arguments{
\item{freq}{}
}
\value{

}
\description{
Get progress per time interval
}
