% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_api.R
\name{read_delorme}
\alias{read_delorme}
\title{Read data from DeLorme inReach API}
\usage{
read_delorme(id, date1, date2 = NULL)
}
\arguments{
\item{id}{inReach ID}

\item{date1}{Start date}

\item{date2}{End date}
}
\value{
A data.frame with three columns: lat (latitude), lon (longitude),
 time (timestamp)
}
\description{
Read data from DeLorme inReach API
}
\references{
\url{https://support.delorme.com/kb/articles/26-about-inreach-kml-feeds}
}

