% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_API.R
\name{read_inreach}
\alias{read_inreach}
\alias{read_delorme}
\title{Read data from Garmin inReach API}
\usage{
read_inreach(id, date1, date2 = NULL)

read_delorme(id, date1, date2 = NULL)
}
\arguments{
\item{id}{inReach ID}

\item{date1}{Start date}

\item{date2}{End date}
}
\value{
A data.frame with three columns: lat (latitude), lon (longitude),
 time (timestamp)
}
\description{
Read data from Garmin inReach API
}
\references{
\url{https://support.garmin.com/en-US/?faq=tdlDCyo1fJ5UxjUbA9rMY8}
}
