% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_API.R
\name{read_MT_hist}
\alias{read_MT_hist}
\title{Read from MarineTraffic historical positions API}
\usage{
read_MT_hist(api_key, mmsi, from_date, outfile)
}
\arguments{
\item{from_date}{}
}
\value{

}
\description{
Read from MarineTraffic historical positions API
}
