% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_api.R
\name{read_spot}
\alias{read_spot}
\title{Read data from the SPOT API}
\usage{
read_spot(id, all = FALSE, password = NULL)
}
\arguments{
\item{id}{SPOT feed ID}

\item{all}{Fetch most recent 50 trackpoint (default) or all available}

\item{password}{SPOT feed password}
}
\value{
A data.frame with three columns: lat (latitude), lon (longitude),
 time (timestamp)
}
\description{
Read data from the SPOT API
}
\references{
\url{http://faq.findmespot.com/index.php?action=showEntry&data=69}
}
