% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROM.R
\name{reduceStdA}
\alias{reduceStdA}
\title{Compute reduced standardized coefficient matrix: (E^-1 V)^T A V}
\usage{
reduceStdA(StdV, A, V)
}
\description{
Compute reduced standardized coefficient matrix: (E^-1 V)^T A V
}
