% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{RiemannianDist}
\alias{RiemannianDist}
\title{The Riemannian distance between two subspaces: 2-norm of principal angles}
\usage{
RiemannianDist(X, Y, normalize = TRUE)
}
\description{
The Riemannian distance between two subspaces: 2-norm of principal angles
}
