% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_speed.R
\name{route_dist}
\alias{route_dist}
\title{Calculate route distance}
\usage{
route_dist(route)
}
\arguments{
\item{route}{df with 2 cols, lat and lon}
}
\value{

}
\description{
Calculate route distance
}
