% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{secs_to_HMS}
\alias{secs_to_HMS}
\title{Convert seconds to hours:minutes:seconds}
\usage{
secs_to_HMS(seconds)
}
\arguments{
\item{seconds}{}
}
\value{
Vector of c(hours, minutes, seconds)
}
\description{
Convert seconds to hours:minutes:seconds
}

