% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot.R
\name{spot_build}
\alias{spot_build}
\title{Read multiple pages of SPOT data}
\usage{
spot_build(id, all = FALSE, password = NULL)
}
\arguments{
\item{password}{}
}
\value{

}
\description{
Read multiple pages of SPOT data
}
