% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot.R
\name{spot_cols}
\alias{spot_cols}
\title{Convert raw SPOT messages to lat/lon/time data-frame}
\usage{
spot_cols(msgs)
}
\arguments{
\item{msgs}{}
}
\value{

}
\description{
Convert raw SPOT messages to lat/lon/time data-frame
}
