% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot.R
\name{spot_colsDT}
\alias{spot_colsDT}
\title{Convert raw SPOT messages to lat/lon/time data-frame (DT implementation)}
\usage{
spot_colsDT(msgs)
}
\arguments{
\item{msgs}{}
}
\description{
Convert raw SPOT messages to lat/lon/time data-frame (DT implementation)
}
