% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_API.R
\name{spot_page}
\alias{spot_page}
\title{Read single page of data from the SPOT API}
\usage{
spot_page(url)
}
\arguments{
\item{url}{SPOT API URL}
}
\value{
A data.frame with three columns: lat (latitude), lon (longitude),
 time (timestamp)
}
\description{
Read single page of data from the SPOT API
}
