% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolate-Subspace.R
\name{SubspaceInterp}
\alias{SubspaceInterp}
\title{Subspace interpolation}
\usage{
SubspaceInterp(theta, nr = l, method = "Lagrange")
}
\arguments{
\item{theta}{New parameter point}

\item{nr}{Number of (parameter, reduced basis) pairs used in the interpolation.
Defaults to all observations.}

\item{method}{Interpolation scheme, either "Lagrange" or "RBF" (for multiquadrics)}
}
\value{
An orthonormal basis for the interpolated subspace.
}
\description{
Subspace interpolation
}
\details{
It not clear from \link{@Amsallem2008} and later papers that
whether the reference point should be one of the interpolation points,
but it is more reasonable to do than exclude it.
The tangent vector of the reference point is identically zero.
}
\note{
Require (thetaTrain, listXtrain)
}
\references{
@Amsallem2008
}
