% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolate-Subspace.R
\name{TangentInterpLagrange}
\alias{TangentInterpLagrange}
\title{Tangent vector Lagrange interpolation (one-parameter)}
\usage{
TangentInterpLagrange(theta, ref, neigh)
}
\arguments{
\item{theta}{New parameter point, a scalar}

\item{ref}{Index of the reference training parameter point}

\item{neigh}{Indices of the used neighboring training parameter points}
}
\value{
Tangent vector
}
\description{
Tangent vector Lagrange interpolation (one-parameter)
}
\note{
Require (thetaTrain, listGL)
}
