% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolate-Subspace.R
\name{TangentInterpRBF}
\alias{TangentInterpRBF}
\title{Tangent vector interpolation with (multiquadric) raidal basis function}
\usage{
TangentInterpRBF(theta, ref, neigh)
}
\arguments{
\item{theta}{New parameter point, a vector}
}
\description{
Tangent vector interpolation with (multiquadric) raidal basis function
}
\note{
Require (thetaTrain, listGL, listXtrain)
}
