% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thermal}
\alias{thermal}
\title{System matrices for thermal model of a micropyros thruster.}
\format{
A list:
\describe{
\item{BCIE}{diagonal mass matrix}
\item{BCIA}{symmetric sparse coefficient matrix}
\item{BCIAtop, BCIAbottom, BCIAside}{diagonal coefficient matrices
for the top/bottom/side boundary conditions}
\item{BCIB}{input 1-column sparse matrix}
\item{BCIC}{output 7-row (index) matrix}
}
}
\source{
\url{https://morwiki.mpi-magdeburg.mpg.de/morwiki/index.php/Thermal_Model}
}
\usage{
thermal
}
\description{
System dimension 4,257, with one input and seven output; depending on three parameters.
}
\keyword{datasets}
