% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_speed.R
\name{time_remain}
\alias{time_remain}
\title{Estimated time remaining to finish}
\usage{
time_remain(track_mut, meters_remain)
}
\arguments{
\item{track_mut}{via trk_mutate()}

\item{meters_remain}{to finish}
}
\value{
dbl seconds
}
\description{
Estimated time remaining to finish
}
