% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_speed.R
\name{trk_dist}
\alias{trk_dist}
\title{Total track distance (sum of trackpoints)}
\usage{
trk_dist(track_mut)
}
\arguments{
\item{mut_track}{mutated track (via trk_mutate)}
}
\value{
meters
}
\description{
Total track distance (sum of trackpoints)
}
