% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_munge.R
\name{trk_downsample}
\alias{trk_downsample}
\title{Downsample a track with evenly spaced times to a target number of rows}
\usage{
trk_downsample(trk_data, target_n)
}
\arguments{
\item{trk_data}{}
}
\description{
Downsample a track with evenly spaced times to a target number of rows
}
