% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{trk_map}
\alias{trk_map}
\title{Create a map of a track}
\usage{
trk_map(track, freq = "60 min", tz_offset = 0, provider_tiles = NULL,
  custom_tiles = NULL, p_tile_opts = list(), c_tile_opts = list(),
  p_tile_labs = NULL, c_tile_labs = NULL, labels = NULL,
  circleColor = NULL, lineColor = NULL)
}
\arguments{
\item{track}{track data}

\item{freq}{frequency for trk_reduce}

\item{tz_offset}{hours offset from GMT}

\item{lineColor}{}
}
\description{
Create a map of a track
}
