% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_speed.R
\name{trk_mutate}
\alias{trk_mutate}
\title{Mutate a track
Add derived variables: distance from previous trackpoint,
elapsed time since last trackpoint, speed (kph), bearing.}
\usage{
trk_mutate(track)
}
\arguments{
\item{track}{}
}
\value{
data.frame
}
\description{
Mutate a track
Add derived variables: distance from previous trackpoint,
elapsed time since last trackpoint, speed (kph), bearing.
}
