% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_speed.R
\name{trk_mutate_dplyr}
\alias{trk_mutate_dplyr}
\title{Mutate a track (dplyr implementation)
Add derived variables: distance from previous trackpoint,
elapsed time since last trackpoint, speed (kph), bearing.}
\usage{
trk_mutate_dplyr(track)
}
\arguments{
\item{track}{a data.frame with 3 cols: 'lat', 'lon', 'time'}
}
\value{
a data.frame with 4 additional cols
}
\description{
Mutate a track (dplyr implementation)
Add derived variables: distance from previous trackpoint,
elapsed time since last trackpoint, speed (kph), bearing.
}
