% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{trk_since_last}
\alias{trk_since_last}
\title{Elapsed time since last trackpoint}
\usage{
trk_since_last(track, units)
}
\arguments{
\item{units}{}
}
\value{
difftime
}
\description{
Elapsed time since last trackpoint
}
