% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{base_map}
\alias{base_map}
\title{Create a blank map}
\usage{
base_map(provider_tiles = NULL, custom_tiles = NULL,
  p_tile_opts = list(), c_tile_opts = list(), p_tile_labs = NULL,
  c_tile_labs = NULL, proxy_map_id = NULL)
}
\arguments{
\item{provider_tiles}{character vector of known map tile providers}

\item{custom_tiles}{character vector of URLs of custom tiles}

\item{p_tile_opts}{list of providerTileOptions for the provider tiles}

\item{c_tile_opts}{list of tileOptions for the custom tiles}

\item{p_tile_labs}{character vector of labels for the provider tiles}

\item{c_tile_labs}{character vector of labels for the custom tiles}

\item{proxy_map_id}{map ID of a leaflet map if called from a Shiny app}
}
\value{
map a leaflet map
}
\description{
Create a blank map
}
