% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_speed.R
\name{dist_from_start}
\alias{dist_from_start}
\title{Distance from start, in a straight line}
\usage{
dist_from_start(track, route = NULL)
}
\arguments{
\item{route}{}
}
\value{
meters
}
\description{
Distance from start, in a straight line
}
