% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_path_lines}
\alias{map_path_lines}
\title{Map Path Lines}
\usage{
map_path_lines(map, data, legendGroup = NULL, lineColor = "#FF4900",
  ...)
}
\arguments{
\item{...}{}
}
\value{
map
}
\description{
Map Path Lines
}
