% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_path_points}
\alias{map_path_points}
\title{Map Path Points}
\usage{
map_path_points(map, data, legendGroup = NULL, circleColor = "#FF4900",
  circleRadius = 50, ...)
}
\arguments{
\item{...}{}
}
\value{
map
}
\description{
Map Path Points
}
