% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_parse.R
\name{read_spot_csv}
\alias{read_spot_csv}
\title{Read SPOT-exported track in CSV format}
\usage{
read_spot_csv(csvFile)
}
\arguments{
\item{csv_file}{}
}
\value{
data.frame
}
\description{
Read SPOT-exported track in CSV format
}

