% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_speed.R
\name{time_remain}
\alias{time_remain}
\title{Estimated time (in seconds) remaining to finish}
\usage{
time_remain(track_mut, meters_to_fin)
}
\arguments{
\item{meters_to_fin}{}
}
\value{
seconds
}
\description{
Estimated time (in seconds) remaining to finish
}

