% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_speed.R
\name{trk_dist}
\alias{trk_dist}
\title{Total track distance
Sum of distances between trackpoints}
\usage{
trk_dist(track = NULL, track_mut = NULL)
}
\arguments{
\item{mut_track}{}
}
\value{
meters
}
\description{
Total track distance
Sum of distances between trackpoints
}
