% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{trk_map}
\alias{trk_map}
\title{Create a map of a track}
\usage{
trk_map(
  track,
  freq = NULL,
  tz_offset = 0,
  provider_tiles = NULL,
  custom_tiles = NULL,
  p_tile_opts = list(),
  c_tile_opts = list(),
  p_tile_labs = NULL,
  c_tile_labs = NULL,
  labels = NULL,
  circleColor = NULL,
  lineColor = NULL
)
}
\arguments{
\item{track}{track data}

\item{freq}{frequency for trk_reduce}

\item{tz_offset}{hours offset from GMT}

\item{provider_tiles}{character vector of known map tile providers}

\item{custom_tiles}{character vector of URLs of custom tiles}

\item{p_tile_opts}{list of providerTileOptions for the provider tiles}

\item{c_tile_opts}{list of tileOptions for the custom tiles}

\item{p_tile_labs}{character vector of labels for the provider tiles}

\item{c_tile_labs}{character vector of labels for the custom tiles}

\item{labels}{labels for each trackpoint}

\item{circleColor}{color of each trackpoint}

\item{lineColor}{color line connecting the trackpoints}
}
\description{
Create a map of a track
}
